pyproject.toml
[project]
name = "robotics-school-site"
version = "0.1.0"
description = ""
authors = [
    {name = "moreveal", email = "morrisreveal@gmail.com"},
    {name = "d3ede"}
]
requires-python = ">=3.11"
dependencies = [
    "django (==5.2.12)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "django-ckeditor (>=6.7.3,<7.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[dependency-groups]
dev = [
    "watchdog (>=6.0.0,<7.0.0)"
]
